fixup parcel integration template
