+++ integration/helpers/rsc-parcel/src/parcel-entry-wrapper.ts
@@ -1,0 +1,5 @@
+"use server-entry";
+
+import "./browser";
+
+export function assets() {}
